Working

+++ app/src/main/java/com/example/riot/data/Match.kt
@@ -77,7 +77,7 @@ data class SessionPlaytime(
 data class Behavior(
     val afk_rounds: Double,
     val friendly_fire: FriendlyFire,
-    val rounds_in_spawn: Int
+    val rounds_in_spawn: Float
 ) : Serializable
 
 data class FriendlyFire(
